+++ src/components/burger-constructor/burger-constructor.module.scss
@@ -24,20 +24,6 @@
 	//height: 100%;
 }
 
-.item {
-	display: flex;
-	align-items: center;
-
-	&:not(:last-child) {
-		margin-bottom: 16px;
-	}
-
-	&__drag {
-		font-size: 0;
-		cursor: pointer;
-	}
-}
-
 .bun {
 	padding-right: 16px;
 }
