+++ src/components/burger-ingredients/burger-ingredients.module.scss
@@ -26,10 +26,6 @@
 		}
 	}
 
-	&__tabs {
-		display: flex;
-	}
-
 	&__section {
 		max-height: 630px;
 		margin-bottom: 52px;
