+++ src/components/tabs/tabs.module.scss
@@ -1,0 +1,3 @@
+.tabs {
+	display: flex;
+}
